Cleanup BlendShapeMerger. Removed BlendShapeMerger.Clear
