+++ Scripts/BlendShape/BlendShapeBindingMerger.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ca4c8446451eeed46b1598db9e08bb73
+timeCreated: 1541229189
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
