+++ Scripts/BlendShape/MaterialValueBindingMerger.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 4ccc88d77d5d1e74499d053083ade08d
+timeCreated: 1541229189
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
